+++ lib/dartapi_db.dart
@@ -1,10 +1,12 @@
 library;
 
 export 'core/db_result.dart';
+export 'core/db_exception.dart';
 export 'core/dartapi_db_core.dart';
 export 'core/db_transaction.dart';
 export 'core/query_extension.dart';
 export 'query/query_builder.dart';
+export 'query/db_page.dart';
 
 export 'drivers/postgres/postgres_database.dart';
 export 'drivers/mysql/mysql_database.dart';